% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_description.R
\docType{data}
\name{comadre_podoc}
\alias{comadre_podoc}
\title{Copy from the COMADRE database.}
\format{
\code{comadre_podoc}
Is a Rcompadre object with a COM(P)ADRE database ('CompadreDB') including 2 SPECIES and 2 MATRICES.
}
\source{
\url{https://compadre-db.org/Data/Comadre}
}
\usage{
comadre_podoc
}
\description{
An example database with matrix population models (MPMs) of two South American turtles species.
Used for development and testing. The MPMs consist stage classes.
}
\details{
Copy from the COMADRE database with Podocnemididae.
}
\examples{
\dontrun{
# look at what is in the database
comadre_podoc
}
}
\references{
Jones, O. R., et. al. (2022).
Rcompadre and rage—Two R packages to facilitate the use of the
COMPADRE and COMADRE databases and calculation of life‐history
traits from matrix population models.
Methods in Ecology and Evolution, 13(4), 770-781.
}
\keyword{datasets}
