% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_beta.R
\name{make_beta}
\alias{make_beta}
\title{Beta distribution around known survival values.}
\usage{
make_beta(
  survival_mean = NA,
  survival_min = NA,
  survival_max = NA,
  alpha_start = 40,
  beta_start = 2
)
}
\arguments{
\item{survival_mean}{Numeric value. Mean desired survival.}

\item{survival_min}{Numeric value. Minimum survival.}

\item{survival_max}{Numeric value. Maximum survival.}

\item{alpha_start}{Numeric value. Alpha parameter, the first shape parameter of the beta distribution.}

\item{beta_start}{Numeric value. Beta parameter, the second shape parameter of the beta distribution.}
}
\value{
Creates a data.frame with survival values and beta distribution parameters".
}
\description{
Beta distribution around known survival values.
}
\examples{
\dontrun{
# Here symmetric mean = 30 Y, min = 20 Y and max = 70 Y.
adult_survival_symmetric <- TACAR::make_beta(survival_mean = 0.95244,
                                     survival_min = 0.9215,
                                      survival_max = 0.98153,
                                      alpha_start = 200,
                                     beta_start = 200)
}
}
