% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_half_normal.R
\name{make_half_normal}
\alias{make_half_normal}
\title{Half-normal distribution around known survival values.}
\usage{
make_half_normal(
  survival_mean = NA,
  survival_min = NA,
  survival_max = NA,
  n_samples = 1000,
  sigma_start = NA,
  optimize_sigma = TRUE
)
}
\arguments{
\item{survival_mean}{Numeric value. Mean desired survival.}

\item{survival_min}{Numeric value. Minimum survival.}

\item{survival_max}{Numeric value. Maximum survival.}

\item{n_samples}{Number of samples. Default 1000.}

\item{sigma_start}{Positive valued scale parameter.}

\item{optimize_sigma}{Logical. Optimize sigma value selection. Default TRUE.}
}
\value{
Creates a data.frame with survival values following the half-normal distribution.
}
\description{
Half-normal distribution around known survival values.
}
\examples{
\dontrun{
survival_values_hn <- make_half_normal(survival_mean = 0.95244,
survival_min = 0.9215,
survival_max = 0.98153)
hist(survival_values_hn$survival_values)
mean(survival_values_hn$survival_values)
}
}
