% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop03_doproj_stoch.R
\name{pop03_doproj_stoch}
\alias{pop03_doproj_stoch}
\title{Create data.frame with stochastic projection results.}
\usage{
pop03_doproj_stoch(x)
}
\arguments{
\item{x}{data.frame created by "PopPrep.R".}
}
\value{
Creates a data.frame with projection results.
}
\description{
Creates data.frame with projection results for use in subsequent
population projection function ("PopProj.R").
}
\details{
This function applies stochastic population projection moodels.
}
\examples{
\dontrun{
dfproj_vals <- pop03_doproj_stoch()
}

}
